+++ I_OTaxis/src/com/test/iotax/vehicle/CargoTaxiImpl.java
@@ -9,12 +9,12 @@ public class CargoTaxiImpl implements IVehicle {
 	 private String model;
 	 private String color;
 	 private int yearProduction;
-	 private float price;
-	 private float fuelConsumption;
+	 private int price;
+	 private int fuelConsumption;
 	 private int maxSpeed;
 	 private int load—apacity;
 	
-	public CargoTaxiImpl(int id, String mark, String model, String color, int yearProduction, float price, float fuelConsumption, int maxSpeed, int loadCapacity) {
+	public CargoTaxiImpl(int id, String mark, String model, String color, int yearProduction, int price, int fuelConsumption, int maxSpeed, int loadCapacity) {
 		this.id = id;
 		this.mark = mark;
 		this.model = model;
@@ -70,7 +70,7 @@ public class CargoTaxiImpl implements IVehicle {
 		return price;
 	}
 
-	public void setPrice(float price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
 
@@ -78,7 +78,7 @@ public class CargoTaxiImpl implements IVehicle {
 		return fuelConsumption;
 	}
 
-	public void setFuelConsumption(float fuelConsumption) {
+	public void setFuelConsumption(int fuelConsumption) {
 		this.fuelConsumption = fuelConsumption;
 	}
